requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
transformers==4.46.3
torch==2.5.1
pillow==11.0.0
requests==2.32.3
motor==3.6.0
python-dotenv==1.0.1
